+++ sdk/core/src/main/java/com/google/cloud/trace/core/AbstractTraceContextHandler.java
@@ -12,9 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.cloud.trace;
+package com.google.cloud.trace.core;
 
-import com.google.cloud.trace.core.TraceContext;
 import java.util.ArrayDeque;
 import java.util.Collections;
 import java.util.logging.Logger;
