+++ sdk/core/src/main/java/com/google/cloud/trace/core/TraceContextHandler.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.cloud.trace;
+package com.google.cloud.trace.core;
 
 import com.google.cloud.trace.core.TraceContext;
